Move textual error handling to the message base class
