+++ tests/rabbit/messages/test_update_sample_message.py
@@ -33,11 +33,6 @@ def test_has_errors_is_true_after_feedback_error_logged(subject):
     assert subject.has_errors is True
 
 
-def test_has_errors_is_true_after_textual_error_logged(subject):
-    subject._textual_errors.append(MagicMock())
-    assert subject.has_errors is True
-
-
 def test_message_uuid_gives_expected_value(subject):
     assert subject.message_uuid.name == FIELD_MESSAGE_UUID
     assert subject.message_uuid.value == "UPDATE_SAMPLE_MESSAGE_UUID"
@@ -110,31 +105,3 @@ def test_feedback_errors_list_is_immutable(subject):
     errors.remove(errors[0])
     assert len(errors) == 0
     assert len(subject.feedback_errors) == 1  # Hasn't been modified
-
-
-@pytest.mark.parametrize(
-    "errors, headline",
-    [
-        [[], "No errors were reported during processing."],
-        [["Error 1"], "1 error was reported during processing."],
-        [["Error 1", "Error 2"], "2 errors were reported during processing."],
-        [["Error 1", "Error 2", "Error 3"], "3 errors were reported during processing."],
-        [["Error 1", "Error 2", "Error 3", "Error 4"], "4 errors were reported during processing."],
-        [["Error 1", "Error 2", "Error 3", "Error 4", "Error 5"], "5 errors were reported during processing."],
-    ],
-)
-def test_textual_errors_summary_is_accurate_for_up_to_5_errors(subject, errors, headline):
-    subject._textual_errors = errors
-    assert subject.textual_errors_summary == [headline] + errors
-
-
-def test_textual_errors_summary_is_accurate_for_6_errors(subject):
-    subject._textual_errors = ["Error 1", "Error 2", "Error 3", "Error 4", "Error 5", "Error 6"]
-    assert subject.textual_errors_summary == [
-        "6 errors were reported during processing. Only the first 5 are shown.",
-        "Error 1",
-        "Error 2",
-        "Error 3",
-        "Error 4",
-        "Error 5",
-    ]
